🚧 Port new line dir functions
